# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Pinterest Web Clone API",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:5000" }],
  "paths": {
    "/users": {
      "post": {
        "tags": ["Users"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "user1"
                  },
                  "email": {
                    "type": "string",
                    "example": "user1@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password1"
                  },
                  "name": {
                    "type": "string",
                    "example": "User 1"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "409": {
            "description": "conflict"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Users"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "user1"
                  },
                  "password": {
                    "type": "string",
                    "example": "password1"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/boards": {
      "post": {
        "tags": ["Boards"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Example"
                  },
                  "is_private": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Boards"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/boards/{board_id}": {
      "get": {
        "tags": ["Boards"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "delete": {
        "tags": ["Boards"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/pins": {
      "post": {
        "tags": ["Pins"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Example"
                  },
                  "image_url": {
                    "type": "string",
                    "example": "https://example.com"
                  },
                  "board_id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Pins"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/pins/{pin_id}": {
      "get": {
        "tags": ["Pins"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "pin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "delete": {
        "tags": ["Pins"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "pin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}